+++ app/src/main/java/ru/netology/diploma/adapter/posts/PostRemoteMediator.kt
@@ -8,10 +8,15 @@ import ru.netology.diploma.db.AppDb
 import ru.netology.diploma.entity.PostEntity
 import ru.netology.diploma.entity.PostKeyEntry
 import ru.netology.diploma.error.ApiError
+import ru.netology.diploma.repository.AppNetState
+import ru.netology.diploma.repository.AuthMethods
+
 
 @ExperimentalPagingApi
 class PostRemoteMediator(private val api: ApiService,
-                         private val base: AppDb)
+                         private val base: AppDb,
+                         private val repoNetwork: AuthMethods
+)
     : RemoteMediator<Int, PostEntity>() {
 
     override suspend fun load(
@@ -19,13 +24,16 @@ class PostRemoteMediator(private val api: ApiService,
         state: PagingState<Int, PostEntity>
             ): MediatorResult {
         try {
-            val response = when (loadType){
-                else -> {
-                    base.postDao().deleteAll()
-                    //api.getLatestPosts(state.config.pageSize)
-                    api.getAll()
-                }
-               /*   LoadType.REFRESH -> {
+            val connected = repoNetwork.checkConnection() == AppNetState.CONNECTION_ESTABLISHED
+
+            if (connected) {
+                val response = when (loadType) {
+                    else -> {
+                        base.postDao().deleteAll()
+                        //api.getLatestPosts(state.config.pageSize)
+                        api.getAll()
+                    }
+                    /*   LoadType.REFRESH -> {
                       base.postDao().deleteAll()
                       api.getLatestPosts(state.config.pageSize)
                   }
@@ -37,71 +45,74 @@ class PostRemoteMediator(private val api: ApiService,
                       val id = base.keyWorkDao().max() ?: return MediatorResult.Success(false)
                     //  api.getAfterPosts(id, state.config.pageSize)
                   }*/
-            }
-
-           if (! response.isSuccessful){
-               Log.e("OkHttpClient", "remote mediator ApiError")
-               throw ApiError(response.code(), response.message())
-           }
-
-            val body = response.body() ?: throw  ApiError(
-                response.code(),
-                response.message()
-            )
-
-            if (body.isEmpty()){
-                Log.e("OkHttpClient", "remote mediator success")
-                return MediatorResult.Success(true)
-
-            }
-
-            base.withTransaction {
-                when (loadType){
-                    LoadType.REFRESH -> {
-                        base.keyPostPaginationDao().insert(
-                            listOf(
-                                 PostKeyEntry(
-                                     PostKeyEntry.Type.PREPEND,
-                                     body.first().id
-                                 ),
-                                PostKeyEntry(
-                                    PostKeyEntry.Type.APPEND,
-                                    body.last().id
-                                )
-                            )
-                        )
-                        base.postDao().deleteAll()
-                    }
-                    LoadType.PREPEND -> {
-                        base.keyPostPaginationDao().insert(
-                            listOf(
-                                PostKeyEntry(
-                                    PostKeyEntry.Type.PREPEND,
-                                    body.first().id
-                                ),
-                            )
-                        )
-
-                    }
-                    LoadType.APPEND ->  {
-                        base.keyPostPaginationDao().insert(
-                            listOf(
-                                PostKeyEntry(
-                                    PostKeyEntry.Type.APPEND,
-                                    body.last().id
-                                )
-                            )
-                        )
-
-                    }
                 }
 
-                base.postDao().insert(body.map(PostEntity.Companion::fromDto))
+                if (!response.isSuccessful) {
+                    Log.e("OkHttpClient", "remote mediator ApiError")
+                    throw ApiError(response.code(), response.message())
+                }
+
+                val body = response.body() ?: throw  ApiError(
+                    response.code(),
+                    response.message()
+                )
+
+                if (body.isEmpty()) {
+                    Log.e("OkHttpClient", "remote mediator success")
+                    return MediatorResult.Success(true)
+
+                }
+
+                base.withTransaction {
+                    when (loadType) {
+                        LoadType.REFRESH -> {
+                            base.keyPostPaginationDao().insert(
+                                listOf(
+                                    PostKeyEntry(
+                                        PostKeyEntry.Type.PREPEND,
+                                        body.first().id
+                                    ),
+                                    PostKeyEntry(
+                                        PostKeyEntry.Type.APPEND,
+                                        body.last().id
+                                    )
+                                )
+                            )
+                            base.postDao().deleteAll()
+                        }
+                        LoadType.PREPEND -> {
+                            base.keyPostPaginationDao().insert(
+                                listOf(
+                                    PostKeyEntry(
+                                        PostKeyEntry.Type.PREPEND,
+                                        body.first().id
+                                    ),
+                                )
+                            )
+
+                        }
+                        LoadType.APPEND -> {
+                            base.keyPostPaginationDao().insert(
+                                listOf(
+                                    PostKeyEntry(
+                                        PostKeyEntry.Type.APPEND,
+                                        body.last().id
+                                    )
+                                )
+                            )
+
+                        }
+                    }
+
+                    base.postDao().insert(body.map(PostEntity.Companion::fromDto))
+                }
+
+
+
+                return MediatorResult.Success(true)
+            } else {
+                return MediatorResult.Success(true)
             }
-
-
-
-            return  MediatorResult.Success(true)
         } catch (e: Exception){
             Log.e("OkHttpClient", "remote mediator Exception")
            return MediatorResult.Error(e)
